put motion configuration into its own file
